move GenerateStats to new file,create class Button
