+++ sourceCode/__main__.py
@@ -1,0 +1,4 @@
+from sourceCode import bot
+
+if __name__ == '__main__':
+    bot.setup()
